# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyrich Battery TR API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Battery {
  id: number;
  modelCode: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  voltage: number | null;
  /** @nullable */
  capacity: number | null;
  /** @nullable */
  cca?: number | null;
  type: string;
  technology: string;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  applications?: string | null;
  active: boolean;
  featured: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface BatteryInput {
  modelCode: string;
  name: string;
  description?: string;
  voltage?: number;
  capacity?: number;
  cca?: number;
  type: string;
  technology: string;
  dimensions?: string;
  weight?: number;
  imageUrl?: string;
  applications?: string;
  active?: boolean;
  featured?: boolean;
  sortOrder?: number;
}

export interface BatteryUpdate {
  modelCode?: string;
  name?: string;
  description?: string;
  voltage?: number;
  capacity?: number;
  cca?: number;
  type?: string;
  technology?: string;
  dimensions?: string;
  weight?: number;
  imageUrl?: string;
  applications?: string;
  active?: boolean;
  featured?: boolean;
  sortOrder?: number;
}

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  /** @nullable */
  linkText?: string | null;
  /** @nullable */
  backgroundColor?: string | null;
  active: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface BannerInput {
  title: string;
  subtitle?: string;
  imageUrl?: string;
  linkUrl?: string;
  linkText?: string;
  backgroundColor?: string;
  active?: boolean;
  sortOrder?: number;
}

export interface BannerUpdate {
  title?: string;
  subtitle?: string;
  imageUrl?: string;
  linkUrl?: string;
  linkText?: string;
  backgroundColor?: string;
  active?: boolean;
  sortOrder?: number;
}

export interface Popup {
  id: number;
  title: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  buttonText?: string | null;
  /** @nullable */
  buttonUrl?: string | null;
  active: boolean;
  showOnce?: boolean;
  delaySeconds?: number;
  createdAt: string;
}

export interface PopupInput {
  title: string;
  content?: string;
  imageUrl?: string;
  buttonText?: string;
  buttonUrl?: string;
  active?: boolean;
  showOnce?: boolean;
  delaySeconds?: number;
}

export interface PopupUpdate {
  title?: string;
  content?: string;
  imageUrl?: string;
  buttonText?: string;
  buttonUrl?: string;
  active?: boolean;
  showOnce?: boolean;
  delaySeconds?: number;
}

export interface AdminStats {
  totalBatteries: number;
  activeBatteries: number;
  activeBanners: number;
  activePopups: number;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  token: string;
}

export type ListBatteriesParams = {
  category?: string;
  featured?: boolean;
};

export type ListBannersParams = {
  active?: boolean;
};

export type ListPopupsParams = {
  active?: boolean;
};