# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyrich Battery TR API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalBatteries: number;
  activeBatteries: number;
  activeBanners: number;
  activePopups: number;
}