# lib/api-zod/src/generated/types/banner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyrich Battery TR API
 * OpenAPI spec version: 0.1.0
 */

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  /** @nullable */
  linkText?: string | null;
  /** @nullable */
  backgroundColor?: string | null;
  active: boolean;
  sortOrder: number;
  createdAt: string;
}