# lib/api-zod/src/generated/types/bannerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyrich Battery TR API
 * OpenAPI spec version: 0.1.0
 */

export interface BannerInput {
  title: string;
  subtitle?: string;
  imageUrl?: string;
  linkUrl?: string;
  linkText?: string;
  backgroundColor?: string;
  active?: boolean;
  sortOrder?: number;
}