# lib/api-zod/src/generated/types/battery.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyrich Battery TR API
 * OpenAPI spec version: 0.1.0
 */

export interface Battery {
  id: number;
  modelCode: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  voltage: number | null;
  /** @nullable */
  capacity: number | null;
  /** @nullable */
  cca?: number | null;
  type: string;
  technology: string;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  applications?: string | null;
  active: boolean;
  featured: boolean;
  sortOrder: number;
  createdAt: string;
}