# lib/api-zod/src/generated/types/batteryUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyrich Battery TR API
 * OpenAPI spec version: 0.1.0
 */

export interface BatteryUpdate {
  modelCode?: string;
  name?: string;
  description?: string;
  voltage?: number;
  capacity?: number;
  cca?: number;
  type?: string;
  technology?: string;
  dimensions?: string;
  weight?: number;
  imageUrl?: string;
  applications?: string;
  active?: boolean;
  featured?: boolean;
  sortOrder?: number;
}