# lib/api-zod/src/generated/types/popup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyrich Battery TR API
 * OpenAPI spec version: 0.1.0
 */

export interface Popup {
  id: number;
  title: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  buttonText?: string | null;
  /** @nullable */
  buttonUrl?: string | null;
  active: boolean;
  showOnce?: boolean;
  delaySeconds?: number;
  createdAt: string;
}