# lib/api-zod/src/generated/types/popupInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyrich Battery TR API
 * OpenAPI spec version: 0.1.0
 */

export interface PopupInput {
  title: string;
  content?: string;
  imageUrl?: string;
  buttonText?: string;
  buttonUrl?: string;
  active?: boolean;
  showOnce?: boolean;
  delaySeconds?: number;
}